# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Lampira Tourism Marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ListingCategory = typeof ListingCategory[keyof typeof ListingCategory];


export const ListingCategory = {
  transportation: 'transportation',
  accommodation: 'accommodation',
  restaurant: 'restaurant',
  tour: 'tour',
  event: 'event',
  guide: 'guide',
  souvenir: 'souvenir',
} as const;

export type ListingStatus = typeof ListingStatus[keyof typeof ListingStatus];


export const ListingStatus = {
  active: 'active',
  inactive: 'inactive',
  pending_review: 'pending_review',
} as const;

/**
 * @nullable
 */
export type ListingMetadata = { [key: string]: unknown } | null;

export interface Listing {
  id: number;
  vendorId: number;
  /** @nullable */
  vendorName?: string | null;
  category: ListingCategory;
  name: string;
  description: string;
  city: string;
  /** @nullable */
  address?: string | null;
  price: number;
  currency: string;
  /** @nullable */
  imageUrl?: string | null;
  images?: string[];
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  /** @nullable */
  avgRating?: number | null;
  reviewCount?: number;
  bookingCount?: number;
  isFeatured?: boolean;
  status: ListingStatus;
  /** @nullable */
  metadata?: ListingMetadata;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export interface ListingPage {
  data: Listing[];
  total: number;
  page: number;
  limit: number;
}

export type ListingInputCategory = typeof ListingInputCategory[keyof typeof ListingInputCategory];


export const ListingInputCategory = {
  transportation: 'transportation',
  accommodation: 'accommodation',
  restaurant: 'restaurant',
  tour: 'tour',
  event: 'event',
  guide: 'guide',
  souvenir: 'souvenir',
} as const;

export type ListingInputMetadata = { [key: string]: unknown };

export interface ListingInput {
  category: ListingInputCategory;
  name: string;
  description: string;
  city: string;
  address?: string;
  price: number;
  currency?: string;
  imageUrl?: string;
  images?: string[];
  latitude?: number;
  longitude?: number;
  metadata?: ListingInputMetadata;
}

export type ListingUpdateStatus = typeof ListingUpdateStatus[keyof typeof ListingUpdateStatus];


export const ListingUpdateStatus = {
  active: 'active',
  inactive: 'inactive',
} as const;

export type ListingUpdateMetadata = { [key: string]: unknown };

export interface ListingUpdate {
  name?: string;
  description?: string;
  city?: string;
  address?: string;
  price?: number;
  imageUrl?: string;
  images?: string[];
  latitude?: number;
  longitude?: number;
  status?: ListingUpdateStatus;
  metadata?: ListingUpdateMetadata;
}

export interface DiscoverStats {
  totalListings: number;
  totalVendors: number;
  totalBookings: number;
  totalDestinations: number;
}

export interface CategoryCount {
  category: string;
  count: number;
  label?: string;
}

export type BookingStatus = typeof BookingStatus[keyof typeof BookingStatus];


export const BookingStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  cancelled: 'cancelled',
  completed: 'completed',
} as const;

export interface Booking {
  id: number;
  listingId: number;
  /** @nullable */
  listingName?: string | null;
  /** @nullable */
  listingImageUrl?: string | null;
  /** @nullable */
  listingCategory?: string | null;
  userId: number;
  /** @nullable */
  userName?: string | null;
  /** @nullable */
  userEmail?: string | null;
  status: BookingStatus;
  totalPrice: number;
  /** @nullable */
  commissionAmount?: number | null;
  guests?: number;
  checkInDate: string;
  /** @nullable */
  checkOutDate?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface BookingInput {
  listingId: number;
  checkInDate: string;
  checkOutDate?: string;
  /** @minimum 1 */
  guests: number;
  notes?: string;
}

export type BookingStatusUpdateStatus = typeof BookingStatusUpdateStatus[keyof typeof BookingStatusUpdateStatus];


export const BookingStatusUpdateStatus = {
  confirmed: 'confirmed',
  cancelled: 'cancelled',
  completed: 'completed',
} as const;

export interface BookingStatusUpdate {
  status: BookingStatusUpdateStatus;
}

export interface Review {
  id: number;
  listingId: number;
  userId: number;
  /** @nullable */
  userName?: string | null;
  /** @nullable */
  userAvatarUrl?: string | null;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  /** @nullable */
  comment?: string | null;
  createdAt: string;
}

export interface ReviewInput {
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  comment?: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  tourist: 'tourist',
  vendor: 'vendor',
  admin: 'admin',
} as const;

export interface VendorProfile {
  businessName?: string;
  businessDescription?: string;
  trustScore?: number;
  isVerified?: boolean;
}

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  phone?: string | null;
  language?: string;
  vendorProfile?: VendorProfile;
  createdAt: string;
}

export type UserRegisterInputRole = typeof UserRegisterInputRole[keyof typeof UserRegisterInputRole];


export const UserRegisterInputRole = {
  tourist: 'tourist',
  vendor: 'vendor',
} as const;

export interface UserRegisterInput {
  email: string;
  /** @minLength 6 */
  password: string;
  name: string;
  role: UserRegisterInputRole;
  phone?: string;
  businessName?: string;
}

export interface UserLoginInput {
  email: string;
  password: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface UserUpdate {
  name?: string;
  phone?: string;
  avatarUrl?: string;
  language?: string;
  businessName?: string;
  businessDescription?: string;
}

export interface StatusCount {
  status: string;
  count: number;
}

export interface MonthlyRevenue {
  month: string;
  revenue: number;
  bookings: number;
}

export interface VendorDashboard {
  totalListings: number;
  totalBookings: number;
  pendingBookings: number;
  totalRevenue: number;
  netRevenue?: number;
  avgRating: number;
  trustScore?: number;
  recentBookings: Booking[];
  bookingsByStatus: StatusCount[];
  monthlyRevenue: MonthlyRevenue[];
}

export interface CategoryEarning {
  category: string;
  gross: number;
  commission: number;
  net: number;
  commissionRate: number;
}

export interface VendorEarnings {
  grossRevenue: number;
  totalCommission: number;
  netRevenue: number;
  byCategory: CategoryEarning[];
}

export type GetListingsParams = {
category?: string;
city?: string;
search?: string;
minPrice?: number;
maxPrice?: number;
rating?: number;
page?: number;
limit?: number;
};

export type GetTrendingListingsParams = {
category?: string;
limit?: number;
};

export type GetBookingsParams = {
status?: GetBookingsStatus;
listingId?: number;
};

export type GetBookingsStatus = typeof GetBookingsStatus[keyof typeof GetBookingsStatus];


export const GetBookingsStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  cancelled: 'cancelled',
  completed: 'completed',
} as const;

export type GetVendorBookingsParams = {
status?: GetVendorBookingsStatus;
};

export type GetVendorBookingsStatus = typeof GetVendorBookingsStatus[keyof typeof GetVendorBookingsStatus];


export const GetVendorBookingsStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  cancelled: 'cancelled',
  completed: 'completed',
} as const;