# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Lampira Tourism Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /listings:
    get:
      operationId: getListings
      tags: [listings]
      summary: List all listings
      parameters:
        - name: category
          in: query
          schema: { type: string }
        - name: city
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
        - name: minPrice
          in: query
          schema: { type: number }
        - name: maxPrice
          in: query
          schema: { type: number }
        - name: rating
          in: query
          schema: { type: number }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 12 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingPage"
    post:
      operationId: createListing
      tags: [listings]
      summary: Create a new listing (vendor)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
  /listings/{id}:
    get:
      operationId: getListing
      tags: [listings]
      summary: Get a single listing
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "404":
          description: Not found
    patch:
      operationId: updateListing
      tags: [listings]
      summary: Update a listing (vendor)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListingUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
    delete:
      operationId: deleteListing
      tags: [listings]
      summary: Delete a listing (vendor)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── DISCOVER / WOW ENDPOINTS ───────────────────────────────────────────────
...
  /bookings:
    get:
      operationId: getBookings
      tags: [bookings]
      summary: List bookings for authenticated user or vendor
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, cancelled, completed]
        - name: listingId
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a new booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get booking detail
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}/status:
    patch:
      operationId: updateBookingStatus
      tags: [bookings]
      summary: Accept or reject a booking (vendor)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingStatusUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"

  # ── REVIEWS ───────────────────────────────────────────────────────────────
  /listings/{id}/reviews:
    get:
      operationId: getListingReviews
      tags: [reviews]
      summary: Get reviews for a listing
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Leave a review for a listing
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # ── USERS ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Listing:
      type: object
      required:
        - id
        - vendorId
        - category
        - name
        - description
        - city
        - price
        - currency
        - status
        - createdAt
      properties:
        id: { type: integer }
        vendorId: { type: integer }
        vendorName: { type: ["string", "null"] }
        category:
          type: string
          enum: [transportation, accommodation, restaurant, tour, event, guide, souvenir]
        name: { type: string }
        description: { type: string }
        city: { type: string }
        address: { type: ["string", "null"] }
        price: { type: number }
        currency: { type: string, default: "IDR" }
        imageUrl: { type: ["string", "null"] }
        images: { type: array, items: { type: string } }
        latitude: { type: ["number", "null"] }
        longitude: { type: ["number", "null"] }
        avgRating: { type: ["number", "null"] }
        reviewCount: { type: integer, default: 0 }
        bookingCount: { type: integer, default: 0 }
        isFeatured: { type: boolean, default: false }
        status:
          type: string
          enum: [active, inactive, pending_review]
        metadata: { type: ["object", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: ["string", "null"], format: date-time }
    ListingPage:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Listing"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
    ListingInput:
      type: object
      required: [category, name, description, city, price]
      properties:
        category:
          type: string
          enum: [transportation, accommodation, restaurant, tour, event, guide, souvenir]
        name: { type: string }
        description: { type: string }
        city: { type: string }
        address: { type: string }
        price: { type: number }
        currency: { type: string, default: "IDR" }
        imageUrl: { type: string }
        images: { type: array, items: { type: string } }
        latitude: { type: number }
        longitude: { type: number }
        metadata: { type: object }
    ListingUpdate:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        city: { type: string }
        address: { type: string }
        price: { type: number }
        imageUrl: { type: string }
        images: { type: array, items: { type: string } }
        latitude: { type: number }
        longitude: { type: number }
        status:
          type: string
          enum: [active, inactive]
        metadata: { type: object }

    # ── DISCOVER ─────────────────────────────────────────────────────────────
...
    Booking:
      type: object
      required:
        - id
        - listingId
        - userId
        - status
        - totalPrice
        - checkInDate
        - createdAt
      properties:
        id: { type: integer }
        listingId: { type: integer }
        listingName: { type: ["string", "null"] }
        listingImageUrl: { type: ["string", "null"] }
        listingCategory: { type: ["string", "null"] }
        userId: { type: integer }
        userName: { type: ["string", "null"] }
        userEmail: { type: ["string", "null"] }
        status:
          type: string
          enum: [pending, confirmed, cancelled, completed]
        totalPrice: { type: number }
        commissionAmount: { type: ["number", "null"] }
        guests: { type: integer, default: 1 }
        checkInDate: { type: string, format: date }
        checkOutDate: { type: ["string", "null"], format: date }
        notes: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    BookingInput:
      type: object
      required: [listingId, checkInDate, guests]
      properties:
        listingId: { type: integer }
        checkInDate: { type: string, format: date }
        checkOutDate: { type: string, format: date }
        guests: { type: integer, minimum: 1 }
        notes: { type: string }
    BookingStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [confirmed, cancelled, completed]

    # ── REVIEW ────────────────────────────────────────────────────────────────
    Review:
      type: object
      required: [id, listingId, userId, rating, createdAt]
      properties:
        id: { type: integer }
        listingId: { type: integer }
        userId: { type: integer }
        userName: { type: ["string", "null"] }
        userAvatarUrl: { type: ["string", "null"] }
        rating: { type: integer, minimum: 1, maximum: 5 }
        comment: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    ReviewInput:
      type: object
      required: [rating]
      properties:
        rating: { type: integer, minimum: 1, maximum: 5 }
        comment: { type: string }

    # ── USER ─────────────────────────────────────────────────────────────────
...