# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Lampira Tourism Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from './bookingStatus';

export interface Booking {
  id: number;
  listingId: number;
  /** @nullable */
  listingName?: string | null;
  /** @nullable */
  listingImageUrl?: string | null;
  /** @nullable */
  listingCategory?: string | null;
  userId: number;
  /** @nullable */
  userName?: string | null;
  /** @nullable */
  userEmail?: string | null;
  status: BookingStatus;
  totalPrice: number;
  /** @nullable */
  commissionAmount?: number | null;
  guests?: number;
  checkInDate: Date;
  /** @nullable */
  checkOutDate?: Date | null;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
}