# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Lampira Tourism Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface BookingInput {
  listingId: number;
  checkInDate: Date;
  checkOutDate?: Date;
  /** @minimum 1 */
  guests: number;
  notes?: string;
}